whatsapp-mcp-server/pyproject.toml
[project]
name = "whatsapp-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "requests>=2.32.3",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.1",
]
